Move client cache to own file and cleanup client instance
